refactor instance nax
