+++ src/Naxxramas/Naxxramas_loader.cpp
@@ -23,6 +23,7 @@ void AddSC_boss_sapphiron_40();
 void AddSC_boss_thaddius_40();
 void AddSC_custom_spells_40();
 void AddSC_custom_creatures_40();
+void AddSC_custom_gameobjects_40();
 
 void AddNaxxramas_Scripts()
 {
@@ -47,4 +48,5 @@ void AddNaxxramas_Scripts()
     AddSC_boss_thaddius_40();
     AddSC_custom_spells_40();
     AddSC_custom_creatures_40();
+    AddSC_custom_gameobjects_40();
 }
