+++ src/Naxxramas/scripts/custom_creatures_40.cpp
@@ -5,6 +5,67 @@
 #include "SpellScript.h"
 #include "naxxramas.h"
 
+class npc_naxx40_area_trigger : public CreatureScript
+{
+private:
+    static bool isAttuned(Player* player)
+    {
+        if (player->GetQuestStatus(NAXX40_ATTUNEMENT_1) == QUEST_STATUS_REWARDED)
+            return true;
+        if (player->GetQuestStatus(NAXX40_ATTUNEMENT_2) == QUEST_STATUS_REWARDED)
+            return true;
+        if (player->GetQuestStatus(NAXX40_ATTUNEMENT_3) == QUEST_STATUS_REWARDED)
+            return true;
+        return false;
+    }
+
+public:
+    npc_naxx40_area_trigger() : CreatureScript("npc_naxx40_area_trigger") {}
+
+    struct npc_naxx40_area_triggerAI: public ScriptedAI
+    {
+        npc_naxx40_area_triggerAI(Creature* creature) : ScriptedAI(creature)
+        {
+            me->SetDisplayId(11686); // Invisible
+        }
+
+        void MoveInLineOfSight(Unit* who) override
+        {
+            if (who && me->GetDistance2d(who) < 5.0f)
+            {
+                if (Player* player = who->ToPlayer())
+                {
+                    if (isAttuned(player))
+                    {
+                        player->SetRaidDifficulty(RAID_DIFFICULTY_10MAN_HEROIC);
+                        player->TeleportTo(533, 3005.51f, -3434.64f, 304.195f, 6.2831f);
+                    }
+                }
+
+            }
+            else if (who && me->GetDistance2d(who) < 20.0f)
+            {
+                if (Player* player = who->ToPlayer())
+                {
+                    if (isAttuned(player))
+                    {
+                        GameObject* door = me->FindNearestGameObject(NAXX_STRATH_GATE, 100.0f);
+                        if (door)
+                        {
+                            door->SetGoState(GO_STATE_ACTIVE);
+                        }
+                    }
+                }
+            }
+        }
+    };
+
+    CreatureAI* GetAI(Creature* creature) const override
+    {
+        return new npc_naxx40_area_triggerAI(creature);
+    }
+};
+
 class boss_heigan_eye_stalk_40 : public CreatureScript
 {
 public:
@@ -108,5 +169,6 @@ public:
 
 void AddSC_custom_creatures_40()
 {
+    new npc_naxx40_area_trigger();
     new boss_heigan_eye_stalk_40();
 }
