+++ src/Naxxramas/scripts/custom_gameobjects_40.cpp
@@ -1,0 +1,52 @@
+
+#include "ScriptMgr.h"
+#include "ScriptedCreature.h"
+#include "SpellAuraEffects.h"
+#include "SpellScript.h"
+#include "naxxramas.h"
+#include "VanillaNaxxramas.h"
+
+class gobject_naxx40_tele : public GameObjectScript
+{
+private:
+    static bool isAttuned(Player* player)
+    {
+        if (player->GetQuestStatus(NAXX40_ATTUNEMENT_1) == QUEST_STATUS_REWARDED)
+            return true;
+        if (player->GetQuestStatus(NAXX40_ATTUNEMENT_2) == QUEST_STATUS_REWARDED)
+            return true;
+        if (player->GetQuestStatus(NAXX40_ATTUNEMENT_3) == QUEST_STATUS_REWARDED)
+            return true;
+        return false;
+    }
+
+public:
+    gobject_naxx40_tele() : GameObjectScript("gobject_naxx40_tele") { }
+
+    struct gobject_naxx40_teleAI: GameObjectAI
+    {
+        explicit gobject_naxx40_teleAI(GameObject* object) : GameObjectAI(object) { };
+
+    };
+
+    GameObjectAI* GetAI(GameObject* object) const override
+    {
+        return new gobject_naxx40_teleAI(object);
+    }
+
+    bool OnGossipHello(Player* player, GameObject* /*go*/) override
+    {
+        if ((!sVanillaNaxxramas->requireNaxxStrath || player->GetQuestStatus(NAXX40_ENTRANCE_FLAG) == QUEST_STATUS_REWARDED) && isAttuned(player))
+        {
+            player->SetRaidDifficulty(RAID_DIFFICULTY_10MAN_HEROIC);
+            player->TeleportTo(533, 3005.51f, -3434.64f, 304.195f, 6.2831f);
+        }
+        return true;
+    }
+};
+
+
+void AddSC_custom_gameobjects_40()
+{
+    new gobject_naxx40_tele();
+}
